Added Stuff

+++ Java/Automatic ascent detection and evaluation/src/data/io/CSVFileWriter.java
@@ -8,11 +8,14 @@ import java.nio.file.StandardOpenOption;
 public class CSVFileWriter {
     public static void saveDataToCsvFile(String csvData, String folderPath, String fileName) {
         Path filePath = Path.of(folderPath, fileName);
-
+        
         try {
             Files.createDirectories(filePath.getParent());
-            Files.writeString(filePath, csvData, StandardOpenOption.CREATE);
-            System.out.println("File created");
+            if(Files.exists(filePath)) System.out.println("File allready exists!");
+            else{
+                Files.writeString(filePath, csvData, StandardOpenOption.CREATE);
+                System.out.println("File created");
+            }
         } catch (IOException e) {
             System.out.println("An error occurred while creating or writing to the file.");
             e.printStackTrace();
